requirements.txt
django
djangorestframework
python-dotenv
dnspython
python-whois
python-nmap
requests
beautifulsoup4
jinja2
openai
